file relocations to match edge
